Move implementation of pr_comment to integrations.github.tasks
